Fixed bug where white space caused incorrect header and body
White space is being used to decide which part of $response is to be in the HTTP header and body. However, if there was any further line breaks in the page being served, these caused anything else in the body to be ignored: in my case, there was a line break before the <body> tag, so only the <head>...</head> part of the HTML was served and a blank page was the result. I added a limit of 2 to explode() in order to make sure that ALL the remainder of the response would be in the body. This fixed the problem immediately.